Render the scene with diffuse reflection.
We're making a few assumptions about the nature of light that will allow us to simplify our model:
We assume all light is white and that we can characterize any light by using a real number that we will call intensity.
We are also ignoring the effect of the atmosphere on the light.
We calculate the total intensity of the light for every point of the scene by adding the intensity of the light at the point
for each light in the scene. Ambient lights are assumed to have a constant intensity.For point and directional lights, the
amount of light contributed to the scene will depend on the angle they hit the surface of the sphere at the point of
intersection. To make calculation easier, we assume the area at the point of intersection can be considered flat.