Render scene with support for shadows.
Reuse trace_ray method to compute the closest intersection between a ray of light and the rest of the scene
in order to figure out what objects would be in the shadow of another one.
The parameters are slightly different:
The ray starts from the point of intersection, instead of the camera.
We don't want objects behind the point of intersection to cast shadows over it, so t_min = 0 or very close.
Directional lights are considered to be very far away, so a very distant object could still cast a shadow over the intersection point, which is why t_max = INFINITY.
In the case of point lights, the light_vector is not constant, but computed and in this case t_max = 1, because we don't want objects too far away from the light to cast shadows over the point of intersection. The ray stops at the point light.